Update Buffs
Buffs should now be functional. This includes debuffs.
Included a Minor Armor Increase buff example. +5 to armor.
Fixed loading of Character Hotkeys from the database.

+++ FishMMO-Unity/Assets/Scripts/Shared/Entity/Buff/Template/BuffAttributeTemplate.cs
@@ -6,7 +6,7 @@ namespace FishMMO.Shared
 	[Serializable]
 	public class BuffAttributeTemplate
 	{
-		public long Value;
+		public int Value;
 		[Tooltip("Character Attribute the buff will apply its values to.")]
 		public CharacterAttributeTemplate Template;
 	}
